# Conflicts:
#	.cirrus.yml
#	app/build.gradle
#	app/src/main/java/com/github/sdp/tarjetakuna/MainActivity.kt

+++ app/src/main/java/com/github/sdp/tarjetakuna/model/MagicCard.kt
@@ -27,7 +27,7 @@ data class MagicCard(
      * The total card mana cost.
      * Example : 7
      */
-    val convertedManaCost: Int = 0,
+    val convertedManaCost: Double = 0.0,
 
     /**
      * The card mana cost.
@@ -106,9 +106,6 @@ data class MagicCard(
      * Converts a MagicCard to a MagicCardEntity.
      */
     fun toDBMagicCard(possession: CardPossession = CardPossession.NONE): DBMagicCard {
-        return DBMagicCard(
-            this,
-            possession
-        )
+        return DBMagicCard(this, possession)
     }
 }
